Add files via upload
This Python script, file_deletion.py, is a utility for automating the removal of files with the .NEF extension within a specified folder. The script uses the os module to traverse through the given directory, identifies files with the .NEF extension, and permanently deletes them. This tool can be useful for managing and cleaning up directories containing Nikon Electronic Format (NEF) files commonly used for raw image data. You can use it for deleting files with the specific need. Use with caution, as it performs irreversible actions.